Make dotfiles work properly on spin
This patch adds distro detection and special cases some things that are
different on Fedora and Ubuntu.

Note that this doesn't ensure _everything_ works on Ubuntu, just the
stuff that I need in order to get the work containers running.

I don't use Ubuntu if given the choice, so I make no promise of support.