Replace Hitomi (Surface Pro) with Fern (Thinkpad X1)
I've replaced my Surface Pro 7 with a Thinkpad X1 Carbon.

After spending far too long trying to live with the frustrations of
running Linux on Microsoft's tablet hardware (A real shame as I dearly
love the Surface Pro hardware), I finally bought a laptop to replace it.

This commit introduces `fern` as a first class node, as I've picked that
name for my new thinkpad.

fern is also running Void Linux, so there have been a lot of changes
here to support bootstrapping a system that is incredibly minimal
compared to Fedora/Ubuntu.

This Commit includes several notable changes including:

- Void Linux uses runit instead of Systemd, and has it's own package
  manager (xbps). I have bumped the mitamae version to a new version
  that includes support for these things.

- `seatd` & `turnstile` are implemented to manage sessions and run user
  services as part of the new void prelude.  This provides a minimal
  replacement to the `logind` and `systemctl --user` parts of systemd.

- `greetd`, with the `gtkgreet` greeter to provide a minimal display
  manager using Wayland that's capable of logging into a Sway session.

The rest of the changes are mainly just QoL improvments - tweak configs,
replace Wofi with Fuzzel for application launching, extract proper
recipes for flatpak and pipewire config, that sort of thing.

Some of the new things introduced (like flatpak and pipewire config) are
Void specific for now, for a couple of reasons: a) most other distros
already have them configured, and b) My experiences with Void so far
have been very positive, and I may consider switching to it on my other
machines in the near future.